+++ Cyotek.Windows.Forms.ColorPicker.Demo/BaseForm.cs
@@ -5,12 +5,13 @@ using System.Windows.Forms;
 namespace Cyotek.Windows.Forms.ColorPicker.Demo
 {
   // Cyotek Color Picker controls library
-  // Copyright © 2013-2017 Cyotek Ltd.
+  // Copyright © 2013-2018 Cyotek Ltd.
   // http://cyotek.com/blog/tag/colorpicker
 
   // Licensed under the MIT License. See license.txt for the full text.
 
   // If you use this code in your applications, donations or attribution are welcome
+  // https://www.paypal.me/cyotek
 
   internal partial class BaseForm : Form
   {
@@ -37,6 +38,13 @@ namespace Cyotek.Windows.Forms.ColorPicker.Demo
       base.OnShown(e);
 
       Cursor.Current = Cursors.Default;
+    }
+
+    protected override void OnLoad(EventArgs e)
+    {
+      base.OnLoad(e);
+
+      base.Font = SystemFonts.MessageBoxFont;
     }
 
     #endregion